--- src/components/shared/Checkbox/Checkbox.module.css
@@ -1,7 +1,0 @@
-.checkboxWrapper {
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  gap: 5px;
-  margin-bottom: 10px;
-}
