--- src/components/shared/ImageInput/ImageInput.module.css
@@ -1,13 +1,0 @@
-.imageInput {
-  margin-bottom: 10px;
-  display: flex;
-  flex-direction: column;
-  gap: 5px;
-}
-
-.preview img {
-  max-width: 150px;
-  max-height: 150px;
-  border-radius: 8px;
-  object-fit: cover;
-}
