--- src/components/shared/Input/Input.module.css
@@ -1,14 +1,0 @@
-.inputWrapper {
-  display: flex;
-  flex-direction: column;
-  gap: 5px;
-  margin-block: 10px;
-}
-
-.label {
-}
-
-.input {
-  padding: 5px;
-  max-width: 300px;
-}
