--- src/components/shared/InputError/InputError.module.css
@@ -1,5 +1,0 @@
-.error {
-  color: #ef4444;
-  min-height: 1rem;
-  font-size: 0.875rem;
-}
