--- src/components/shared/InputError/InputError.tsx
@@ -1,5 +1,0 @@
-import styles from './InputError.module.css';
-
-export default function InputError({ error }: { error?: string }) {
-  return <p className={styles.error}>{error}</p>;
-}
