--- src/components/shared/Radio/Radio.module.css
@@ -1,12 +1,0 @@
-.radio {
-  margin-bottom: 10px;
-}
-
-.radio > p {
-}
-
-.radioWrapper {
-  display: flex;
-  align-items: center;
-  gap: 5px;
-}
