--- src/components/shared/Select/Select.module.css
@@ -1,6 +1,0 @@
-.selectWrapper {
-  display: flex;
-  flex-direction: column;
-  gap: 5px;
-  margin-block: 10px;
-}
