+++ src/components/shared/index.tsx
@@ -1,8 +1,4 @@
-import Checkbox from './Checkbox/Checkbox';
-import ImageInput from './ImageInput/ImageInput';
-import Input from './Input/Input';
 import Modal from './Modal/Modal';
-import Radio from './Radio/Radio';
-import Select from './Select/Select';
+import Columns from './Columns/Columns';
 
-export { Checkbox, ImageInput, Input, Modal, Radio, Select };
+export { Modal, Columns };
